add stdint

+++ workshop/hardware/expressions-and-control-flow/workshop/CodeBlocks/16.c
@@ -1,7 +1,8 @@
 #include <stdio.h>
+#include <stdint.h>
 
 int main() {
-	int v = 426;
+	uint16_t v = 426;
 	// print 'Yeah!' if dividable by 4 but print 'End of program' after regardless
 	return 0;
 }
